Refactor commands out into their own modules.
I felt like index.ts was getting awfully cluttered and decided to clear
it out. Passing a context to every function means we don't need to share
our global variables. Also means it might be easier to test these now by
mocking out parts of the context.
