+++ package.json
@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "build": "npm run compile",
+    "clean": "rm -r build && rm -r node_modules",
     "check-all": "npm run compile && npm run test && npm run lint",
     "compile": "tsc --project ./",
     "test": "LOG_LEVEL=debug ts-mocha --paths --file ./src/tests/init \"**/*.spec*.ts\" --ignore \"node_modules/**/*\"",
@@ -57,6 +58,7 @@
     "typescript": "^4.2.2"
   },
   "_moduleAliases": {
+    "@commands": "build/src/commands",
     "@core": "build/src/core",
     "@formatting": "build/src/formatting",
     "@generator": "build/src/generator",
