+++ tsconfig.json
@@ -45,6 +45,7 @@
     // "moduleResolution": "node",            /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */
     "baseUrl": "./src/",                      /* Base directory to resolve non-absolute module names. */
     "paths": {                                /* A series of entries which re-map imports to lookup locations relative to the 'baseUrl'. */
+        "@commands/*": ["commands/*"],
         "@core/*": ["core/*"],
         "@formatting/*": ["formatting/*"],
         "@generator/*": ["generator/*"],
